+++ packages/openapi3/src/cli/actions/convert/transforms/transforms.ts
@@ -1,13 +1,8 @@
 import { OpenAPI3Document } from "../../../../types.js";
-import {
-  TypeSpecModel,
-  TypeSpecNamespace,
-  TypeSpecOperation,
-  TypeSpecProgram,
-  TypeSpecServiceInfo,
-} from "../interfaces.js";
+import { TypeSpecModel, TypeSpecProgram } from "../interfaces.js";
 import { transformComponentParameters } from "./transform-component-parameters.js";
 import { transformComponentSchemas } from "./transform-component-schemas.js";
+import { transformNamespaces } from "./transform-namespaces.js";
 import { transformAllOperationResponses } from "./transform-operation-responses.js";
 import { transformPaths } from "./transform-paths.js";
 import { transformServiceInfo } from "./transform-service-info.js";
@@ -18,7 +13,7 @@ export function transform(openapi: OpenAPI3Document): TypeSpecProgram {
 
   return {
     serviceInfo: transformServiceInfo(openapi.info),
-    ...populateProgramDeclarations(models, operations),
+    ...transformNamespaces(models, operations),
     augmentations: [],
   };
 }
@@ -34,79 +29,4 @@ function collectModels(document: OpenAPI3Document): TypeSpecModel[] {
   transformAllOperationResponses(models, document);
 
   return models;
-}
-
-function getFileNamespaceName(serviceInfo: TypeSpecServiceInfo): string {
-  return serviceInfo.name.replaceAll(/[^\w^\d_]+/g, "");
-}
-
-type TypeSpecProgramDeclarations = Pick<TypeSpecProgram, "models" | "operations" | "namespaces">;
-function populateProgramDeclarations(
-  models: TypeSpecModel[],
-  operations: TypeSpecOperation[]
-): TypeSpecProgramDeclarations {
-  // There can only be 1 file namespace - so if scopes is empty then entity belongs at root level
-  const programDecs: TypeSpecProgramDeclarations = {
-    models: [],
-    operations: [],
-    namespaces: {},
-  };
-
-  expandModels(programDecs, models);
-  expandOperations(programDecs, operations);
-
-  return programDecs;
-}
-
-function expandModels(programDecs: TypeSpecProgramDeclarations, models: TypeSpecModel[]): void {
-  for (const model of models) {
-    const { scope } = model;
-    const namespace = getNamespace(programDecs, scope) ?? createNamespace(programDecs, scope);
-    namespace.models.push(model);
-  }
-}
-
-function expandOperations(
-  programDecs: TypeSpecProgramDeclarations,
-  operations: TypeSpecOperation[]
-): void {
-  for (const operation of operations) {
-    const { scope } = operation;
-    const namespace = getNamespace(programDecs, scope) ?? createNamespace(programDecs, scope);
-    namespace.operations.push(operation);
-  }
-}
-
-function getNamespace(
-  programDecs: TypeSpecProgramDeclarations,
-  scope: string[]
-): TypeSpecNamespace | undefined {
-  if (!scope.length) return programDecs;
-
-  let namespace: TypeSpecNamespace = programDecs;
-  for (const fragment of scope) {
-    if (!namespace) return;
-    namespace = namespace.namespaces[fragment];
-  }
-
-  return namespace;
-}
-
-function createNamespace(
-  programDecs: TypeSpecProgramDeclarations,
-  scope: string[]
-): TypeSpecNamespace {
-  let namespace: TypeSpecNamespace = programDecs;
-  for (const fragment of scope) {
-    if (!namespace.namespaces[fragment]) {
-      namespace.namespaces[fragment] = {
-        namespaces: {},
-        models: [],
-        operations: [],
-      };
-    }
-    namespace = namespace.namespaces[fragment];
-  }
-
-  return namespace;
 }
