+++ packages/openapi3/src/cli/actions/convert/utils/get-scope-and-name.ts
@@ -1,8 +1,8 @@
-import { formatIdentifier } from "@typespec/compiler";
+import { printIdentifier } from "@typespec/compiler";
 
 type ScopeAndName = { scope: string[]; name: string };
 export function getScopeAndName(originalName: string): ScopeAndName {
-  const path = originalName.split(".").map(formatIdentifier);
+  const path = originalName.split(".").map(printIdentifier);
   const name = path.pop()!;
 
   return { scope: path, name };
